+++ Source/ColonyManagerRedux/Comps/CompProperties_ManagerDatabase.cs
@@ -1,0 +1,12 @@
+// CompProperties_ManagerDatabase.cs
+// Copyright (c) 2025 Alexander Krivács Schrøder
+
+namespace ColonyManagerRedux;
+
+public class CompProperties_ManagerDatabase : CompProperties
+{
+    public CompProperties_ManagerDatabase()
+    {
+        compClass = typeof(CompManagerDatabase);
+    }
+}